requirements.txt
numpy==1.26.0
opencv-contrib-python==4.8.0.76
paho-mqtt==1.6.1
Pillow==10.0.1
python-dotenv==1.0.0
pyttsx3
